% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.semi.R
\name{main_kern}
\alias{main_kern}
\title{Reducing Demographic Disparity for Prediction Models in the Presence of Class Imbalance}
\usage{
main_kern(dat.train.s, dat.train.t, dat.s.list, dat.t.list,nm.study,nm.cov,nm.out,sampling,model,fed,nm.study.rm.fed,N.rose, myseed)
}
\arguments{
\item{dat.train.s}{source group in training sample}

\item{dat.train.t}{target group in training sample}

\item{dat.s.list}{list of source groups}

\item{dat.t.list}{list of training groups}

\item{nm.study}{name of the cohort of interest}

\item{nm.cov}{names of covariates to be included}

\item{sampling}{oversampling or not}

\item{model}{to model to be used}

\item{fed}{federated learning or not}

\item{nm.study.rm.fed}{the name of the study to be removed from federated learning}

\item{N.rose}{number of oversamples of ROSE}

\item{myseed}{random seed}

}
\value{
beta coefficients
}
\description{
This function calculates beta coefficient of the TransBalance risk prediction
}
